Fix edges sometimes passing through nodes
For some input graphs I observed that edges (dummy nodes) were passing
through other nodes. It turns out that the code that tries to find the
average median alignment was broken: we used javascript's sort function,
but sort works based on the string representation of the input and so
yielded non-sensical (and layout breaking) results. This commit ensures
that we sort based on difference in numeric magnitude not by difference
in string represenation.

Fixes #8