Fix self loops in layout
We don't support self loops properly yet, but this change fixes a bug in
the interactive demo that would cause edge rendering to abort as soon as
a self loop was hit. The issue was that the demo expects to find an
e.dagre.points property - but we weren't adding one for self loops.