Fix to nesting graph with nested subgraphs
Suppose sg2 is nested inside of sg1. Then the top node of sg1 should
point to the top node of sg2. Similarly, the bottom node of sg2 should
point to the bottom node of sg1. Prior to this change, we had the top
node of sg1 pointing to sg2 and sg2 pointing to the bottom node of sg1.