Order: Return early in order/index.js if cross count is 0
This is both a bug fix in the order pass, as well as a perf
optimization. If the cross count is already zero from the original
ordering then there's no need to run the cross minimization steps.

1. This fixes a layout bug where some graphs will have an incorrect node
   ordering. This happens because the original ordering is already
   optimal (zero crosses) but the cross minimization in order/index.js
   will permute the nodes several times anyways even though it can't
   decrease the cross count.

2. This is a perf optimization. If cross count gets to 0 then we should
   stop the loop. I noticed that running `make bench` shows an increase
   in the ops/sec for the order phase of around 700 to 840.

Note that this breaks atleast one case of compound graphs (the failing
tests) and I'm not entirely sure why compound graphs are required to go
through the sweepLayer pass in order to get a proper layout. It seems
that the cross minimization should be separated from the compound graph
layout. I did notice that moving the cross minimization check to within
the for-loop (after sweepLayer) fixes the test. That may be one possible
fix and I'll put up a separate PR. However, it does not fix the original
bug that I was trying to fix.